improve efficiency
integrate scrapeParseClean module and call script into main.
